feat/left-align-toolbar-widgets
feat: Align left toolbar widgets to the left

Standardizes all items on the left toolbar to be widgets by wrapping QActions in QToolButtons. This allows the toolbar's layout to manage all items consistently.

Sets the alignment of the toolbar's layout to Qt.AlignLeft to ensure all widgets are aligned to the left side of the toolbar.